requirements.txt
Flask
qrcode
Pillow
openpyxl
psycopg2-binary
gunicorn
python-dotenv
